requirements.txt
streamlit>=1.31.0
llama-index-core>=0.10.11
llama-index-llms-gemini>=0.1.6
llama-index-embeddings-huggingface>=0.4.0
llama-index-readers-file>=0.1.6
google-generativeai>=0.3.2
python-dotenv>=1.0.1
pandas>=2.1.0
openpyxl>=3.1.2
watchdog>=3.0.0 